Extract iCal converter interface and create UI
